asistente-billetes-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
Pillow==10.0.1
gtts==2.3.2
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
pytest==7.4.3
python-magic==0.4.27